Moves api endpoints to api file
